+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 1.0.3 - 2021-07-23
+
+### Changes
+* Made constants in OSGB36 final
+
+### Fixed
+* Use correct elipsoid for OSGB36 to WGS84 [Issue #1](https://github.com/dstl/osgb/issues/1)
+* Added missing documentation and corrected typos
+
 ## Version 1.0.2 - 2021-05-25
 
 ### Changed